pyproject.toml
[project]
name = "carpartsscraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "bs4>=0.0.2",
    "celery>=5.5.3",
    "dotenv>=0.9.9",
    "eventlet>=0.40.0",
    "fastapi>=0.115.12",
    "gevent>=25.5.1",
    "playwright>=1.50.0",
    "pydantic-settings>=2.9.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "statsmodels>=0.14.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.14",
]
